+++ CHANGELOG.md
@@ -12,6 +12,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 * **Security**: in case of vulnerabilities.
 
 ## Unreleased
+### Added
+- - Add `r()`, `g()` and `b()` methods to `WideRgb` for easy access to each channel value.
+
+
+### Removed
+- Remove undocumented `XYZ::srgb` method that both clamped out-of-gamut values and converted
+  directly to a gamma encoded `[u8; 3]`. Obtain the same result with the more explicit
+  `xyz.rgb(Some(RgbSpace::SRGB)).clamp().values()`.
+- Remove conversion directly from `WideRgb` to clamped and gamma encoded `[u8; 3]`. Prefer being
+  more explicit by converting to the `Rgb` type in between with one of the provided conversion
+  methods.
+
 
 ### Changed
 - Make indexing into a `Spectrum` with out of bounds wavelengths cause a panic, instead